Added file via upload
Tried testing current step 3 code. Having issues at line 157, C doesn't allow for variable size arrays unless they are dynamically allocated, which are not allowed for this project. Will have to look into alternative methods to implement the  parseStr() function to extract the necessary data to be able to identify elements of interest in the user input.